refactor(utils): refact do embeddings e pgvector para a classe de utils, para evitar repitir o código no código de search
